% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{camp}
\alias{camp}
\title{Censoring-based analysis of microbiome proportions (CAMP)}
\usage{
camp(X, cov)
}
\arguments{
\item{X}{A size (n x p) matrix, p is the number of taxa}

\item{cov}{A length n vector (in case of only group label) or a size (n x q) data frame, q is the number of covariates (including group)}
}
\value{
A length p vector of p-values, or a size (p x q) data frame of p-values corresponding the covariates
}
\description{
Censoring-based analysis of microbiome proportions (CAMP)
}
\examples{
\dontrun{
seed <- 1
n1 <- 100
n2 <- 100
p <- 1000
signal.prop <- 0.8
sum.const <- signal.prop
a <- 353

camp.dat <- GenData(seed, n1, n2, p, signal.prop, sum.const, a)

camp.res <- camp(camp.dat$OTU, camp.dat$Group)
head(camp.res)
}
}
