% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{GenData}
\alias{GenData}
\title{Generate simulated micrboiome data}
\usage{
GenData(
  seed,
  n1,
  n2,
  p,
  signal.prop,
  sum.const,
  a,
  L1 = 25000,
  U1 = 40000,
  L2 = 25000,
  U2 = 40000
)
}
\arguments{
\item{seed}{An integer, seed used for reproducibility in random number generation}

\item{n1}{An integer, the number of samples in Group 1}

\item{n2}{An integer, the number of samples in Group 2}

\item{p}{An integer, the number of taxa to be generated}

\item{signal.prop}{A number between 0 to 1, indicating the proportion of taxa that are signals (non-null)}

\item{sum.const}{A number between 0 to 1, representing the signal taxa's cumulative relative abundance (i.e., how much should they add up to)}

\item{a}{A positive number for the Dirichlet distribution (controls the dispersion of the abundance)}

\item{L1}{An integer, the minimum library size sampling in Group 1. Default is 25000.}

\item{U1}{An integer, the maximum library size sampling in Group 1. Default is 40000.}

\item{L2}{An integer, the minimum library size sampling in Group 2. Default is 25000.}

\item{U2}{An integer, the maximum library size sampling in Group 2. Default is 40000.}
}
\value{
A list containing two elements:
\item{OTU}{A size ((n1 + n2) x p) matrix, representing the OTU table for both groups combined}
\item{Group}{A length (n1 + n2) vector, indicating the group labels ("Group 1" or "Group 2") for each sample}
}
\description{
Generate simulated micrboiome data
}
\examples{
\dontrun{
seed <- 1
n1 <- 100
n2 <- 100
p <- 1000
signal.prop <- 0.8
sum.const <- signal.prop
a <- 353

camp.dat <- GenData(seed, n1, n2, p, signal.prop, sum.const, a)
head(camp.dat)
}
}
