% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functs.R
\name{generate_glm_data}
\alias{generate_glm_data}
\title{Generate GLM data}
\usage{
generate_glm_data(
  design_matrix,
  coefficients,
  family_object,
  add_intercept = TRUE
)
}
\arguments{
\item{design_matrix}{a design matrix}

\item{coefficients}{coefficients (possibly including an intercept)}

\item{family_object}{a family object}

\item{add_intercept}{logical indicating whether to add an intercept to the design matrix}
}
\value{
a vector of synthetic data generated from the GLM
}
\description{
\code{generate_glm_data()} generates synthetic GLM data.
}
\examples{
# simulate data
n <- 5000
x <- rbinom(n = n, size = 1, prob = 0.1)
z <- MASS::mvrnorm(n = n, mu = c(-0.5, 0.5), Sigma = toeplitz(c(1, 0.5)))
family_object <- poisson()
design_matrix <- cbind(x, z)
y <- generate_glm_data(
  design_matrix = design_matrix,
  coefficients = c(0.6, 0.4, 0.1, 0.3),
  family_object = family_object,
  add_intercept = TRUE
)
}
