% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precomputation_functs.R
\name{run_score_stat_precomputation}
\alias{run_score_stat_precomputation}
\title{Perform score stat precomputation}
\usage{
run_score_stat_precomputation(fitted_glm)
}
\description{
Perform score stat precomputation
}
